ts: Migrate `message_store.js` to TypeScript.
Added type annotations for function params, return values and local
variables. Moved temporary types from `./web/src/types.ts` to
`message_store.ts`. Refactored the imports in other files affected
by types move. Exclude `message_store.ts` in the `js_rules` with
`"pattern": "subject|SUBJECT".

Modified test cases so that the synthetic data objects satisfiy
schema validation.
