+++ web/tests/example5.test.js
@@ -93,6 +93,7 @@ run_test("insert_message", ({override}) => {
         sender_id: isaac.user_id,
         id: 1001,
         content: "example content",
+        flags: [],
         topic: "Foo",
     };
